refactor: consolidate interpreter files (5 → 2)
Merge interpreter_common + interpreter_events into interpreter_support.py.
Inline interpreter_runner.py into ModalInterpreter class body, removing
the _foo_impl delegation pattern. Keep interpreter_protocol.py for typing.
