+++ src/fleet_rlm/runtime/execution/interpreter.py
@@ -17,59 +17,61 @@ enabling bidirectional communication for tool calls and structured output.
 from __future__ import annotations
 
 import asyncio
+import inspect
+import json
+import logging
 import queue
 import threading
+import time
+from contextlib import contextmanager
 from typing import Any, Callable, Iterator, Sequence
 
 import dspy
 import modal
-from dspy.primitives import FinalOutput
+from dspy.primitives import CodeInterpreterError, FinalOutput
 
+from fleet_rlm.runtime.execution.core_driver import sandbox_driver
 from fleet_rlm.runtime.execution.profiles import ExecutionProfile  # noqa: F811
 from fleet_rlm.runtime.tools.llm_tools import LLMQueryMixin
 from fleet_rlm.runtime.tools.modal_volumes import VolumeOpsMixin
 
-from .interpreter_common import (
-    async_enter as _async_enter_impl,
-)
-from .interpreter_common import (
-    async_exit as _async_exit_impl,
-)
-from .interpreter_common import (
+from . import driver_factories
+from .interpreter_support import (
+    InterpreterExecutionEventData,
+    complete_event_data,
+    emit_execution_event,
+    execution_profile_context,
     get_registered_tools,
     initialize_llm_query_state,
     initialize_tool_runtime_state,
     set_registered_tools,
+    start_event_data,
+    summarize_code,
 )
-from .interpreter_common import (
+from .interpreter_support import (
+    async_enter as _async_enter_impl,
+)
+from .interpreter_support import (
+    async_exit as _async_exit_impl,
+)
+from .interpreter_support import (
     sync_enter as _sync_enter_impl,
 )
-from .interpreter_common import (
+from .interpreter_support import (
     sync_exit as _sync_exit_impl,
 )
-from .interpreter_events import emit_execution_event as _emit_execution_event_impl
-from .interpreter_events import summarize_code as _summarize_code_impl
-from .interpreter_runner import (
-    _aresolve_app_impl,
-    _ashutdown_impl,
-    _astart_impl,
-    _build_default_image,
-    _build_driver_command_and_sandbox_kwargs_impl,
-    _drain_or_flush_stdin_impl,
-    _execute_impl,
-    _execution_profile_impl,
-    _is_recoverable_exec_channel_error_impl,
-    _is_recoverable_start_error_impl,
-    _module_source_for_sandbox_impl,
-    _output_names_impl,
-    _resolve_app_impl,
-    _shutdown_impl,
-    _start_impl,
-    _start_stdout_reader_impl,
-    _summarize_stdout_impl,
-    _tool_names_impl,
-    _write_line_impl,
-)
+from .output_utils import _redact_sensitive_text, _summarize_stdout
+
+logger = logging.getLogger(__name__)
+
+
+def _build_default_image(
+    *, python_version: str, pip_packages: Sequence[str]
+) -> modal.Image:
+    """Build a default Modal image for sandbox execution."""
+    return modal.Image.debian_slim(python_version=python_version).pip_install(
+        *pip_packages
+    )
 
 
 class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
@@ -203,15 +205,30 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
         self._stdout_reader_thread: threading.Thread | None = None
         self._stderr_iter: Iterator[str] | None = None
 
+    # ------------------------------------------------------------------
+    # Tool property
+    # ------------------------------------------------------------------
+
+    @property
+    def tools(self) -> dict[str, Callable[..., Any]]:
+        """Dictionary of registered tools available to sandboxed code."""
+        return get_registered_tools(self)
+
+    @tools.setter
+    def tools(self, value: dict[str, Callable[..., Any]]) -> None:
+        set_registered_tools(self, value)
+
+    # ------------------------------------------------------------------
+    # Event helpers
+    # ------------------------------------------------------------------
+
     @staticmethod
     def _summarize_code(code: str) -> tuple[str, str]:
         """Return deterministic code hash and compact preview text."""
-        return _summarize_code_impl(code)
+        return summarize_code(code)
 
     def _emit_execution_event(self, payload: dict[str, Any]) -> None:
         """Best-effort execution hook dispatch for observability callbacks."""
-        from .interpreter_events import InterpreterExecutionEventData
-
         event = InterpreterExecutionEventData(
             phase=str(payload.get("phase", "")),
             timestamp=float(payload.get("timestamp", 0.0)),
@@ -227,59 +244,32 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
             error_type=payload.get("error_type"),
             error=payload.get("error"),
         )
-        _emit_execution_event_impl(self, event)
+        emit_execution_event(self, event)
+
+    # ------------------------------------------------------------------
+    # Sandbox I/O
+    # ------------------------------------------------------------------
 
     def _start_stdout_reader(self) -> None:
         """Start a background thread to read sandbox stdout."""
-        _start_stdout_reader_impl(self)
+        if self._stdout_iter is None:
+            return
 
-    @property
-    def tools(self) -> dict[str, Callable[..., Any]]:
-        """Dictionary of registered tools available to sandboxed code."""
-        return get_registered_tools(self)
+        self._stdout_queue = queue.Queue()
+        q = self._stdout_queue
+        iterator = self._stdout_iter
 
-    @tools.setter
-    def tools(self, value: dict[str, Callable[..., Any]]) -> None:
-        set_registered_tools(self, value)
+        def _reader() -> None:
+            try:
+                for line in iterator:
+                    q.put(line)
+            except Exception:
+                logger.exception("Error while reading from sandbox stdout iterator")
+            finally:
+                q.put(None)
 
-    def _resolve_app(self) -> modal.App:
-        """Return a fresh App handle."""
-        return _resolve_app_impl(self)
-
-    async def _aresolve_app(self) -> modal.App:
-        """Return a fresh App handle (async)."""
-        return await _aresolve_app_impl(self)
-
-    @staticmethod
-    def _module_source_for_sandbox(module: Any) -> str:
-        """Return module source with future-import lines stripped for embedding."""
-        return _module_source_for_sandbox_impl(module)
-
-    def _build_driver_command_and_sandbox_kwargs(
-        self, *, app: modal.App
-    ) -> tuple[str, dict[str, Any]]:
-        """Build sandbox driver command and kwargs shared by start/astart."""
-        return _build_driver_command_and_sandbox_kwargs_impl(self, app=app)
-
-    def start(self) -> None:
-        """Start the Modal sandbox and initialize the driver process."""
-        _start_impl(self)
-
-    async def astart(self) -> None:
-        """Start the Modal sandbox and initialize the driver process (async)."""
-        await _astart_impl(self)
-
-    def _tool_names(self) -> list[str]:
-        """Get the list of registered tool names."""
-        return _tool_names_impl(self)
-
-    def _output_names(self) -> list[str]:
-        """Get the list of output field names."""
-        return _output_names_impl(self)
-
-    def _summarize_stdout(self, stdout: str) -> str:
-        """Summarize stdout output to prevent context window pollution."""
-        return _summarize_stdout_impl(self, stdout)
+        self._stdout_reader_thread = threading.Thread(target=_reader, daemon=True)
+        self._stdout_reader_thread.start()
 
     def _drain_or_flush_stdin(self) -> None:
         """Flush sandbox stdin, preferring Modal's async drain when available.
@@ -289,25 +279,276 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
         so bridging to Modal's async stream API with ``asyncio.run(...)`` is safe
         and avoids Modal's AsyncUsageWarning for blocking ``drain()`` calls.
         """
-        _drain_or_flush_stdin_impl(self)
+        if self._stdin is None:
+            raise CodeInterpreterError("Sandbox input stream is not initialized")
+
+        def _drain_or_flush() -> None:
+            drain = getattr(self._stdin, "drain", None)
+            if callable(drain):
+                drain_aio = getattr(drain, "aio", None)
+                if callable(drain_aio):
+                    asyncio.run(drain_aio())
+                    return
+                drain()
+                return
+            flush = getattr(self._stdin, "flush", None)
+            if callable(flush):
+                flush()
+
+        try:
+            asyncio.get_running_loop()
+        except RuntimeError:
+            _drain_or_flush()
+            return
+
+        thread_exc: Exception | None = None
+
+        def _run_async_drain() -> None:
+            nonlocal thread_exc
+            try:
+                _drain_or_flush()
+            except Exception as exc:
+                thread_exc = exc
+
+        drain_thread = threading.Thread(
+            target=_run_async_drain,
+            name="modal-stdin-drain",
+            daemon=True,
+        )
+        drain_thread.start()
+        drain_thread.join()
+        if thread_exc is not None:
+            raise thread_exc
 
     def _write_line(self, payload: dict[str, Any]) -> None:
         """Write a JSON payload to the sandbox stdin."""
-        _write_line_impl(self, payload)
+        if self._stdin is None:
+            raise CodeInterpreterError("Sandbox input stream is not initialized")
+        self._stdin.write(json.dumps(payload) + "\n")
+        self._drain_or_flush_stdin()
+
+    # ------------------------------------------------------------------
+    # App / sandbox resolution
+    # ------------------------------------------------------------------
+
+    def _resolve_app(self) -> modal.App:
+        """Return a fresh App handle."""
+        if self._app_obj is not None:
+            return self._app_obj
+        return modal.App.lookup(self._app_name, create_if_missing=True)
+
+    async def _aresolve_app(self) -> modal.App:
+        """Return a fresh App handle (async)."""
+        if self._app_obj is not None:
+            return self._app_obj
+        return await modal.App.lookup.aio(self._app_name, create_if_missing=True)
+
+    @staticmethod
+    def _module_source_for_sandbox(module: Any) -> str:
+        """Return module source with future-import lines stripped for embedding."""
+        source = inspect.getsource(module)
+        return "\n".join(
+            line
+            for line in source.splitlines()
+            if line.strip() != "from __future__ import annotations"
+        )
+
+    def _build_driver_command_and_sandbox_kwargs(
+        self, *, app: modal.App
+    ) -> tuple[str, dict[str, Any]]:
+        """Build sandbox driver command and kwargs shared by start/astart."""
+        with self._llm_call_lock:
+            self._llm_call_count = 0
+
+        from fleet_rlm.runtime.agent import session_history
+
+        from . import sandbox_assets
+
+        bundled_sources = [
+            self._module_source_for_sandbox(driver_factories),
+            self._module_source_for_sandbox(sandbox_assets),
+            self._module_source_for_sandbox(session_history),
+            inspect.getsource(sandbox_driver),
+            "sandbox_driver()",
+        ]
+        driver_command = "\n\n".join(bundled_sources)
+
+        sandbox_kwargs: dict[str, Any] = {
+            "app": app,
+            "image": self.image,
+            "secrets": self.secrets,
+            "timeout": self.timeout,
+        }
+        if self.idle_timeout is not None:
+            sandbox_kwargs["idle_timeout"] = self.idle_timeout
+        if self.volume_name:
+            self._volume = self._resolve_volume()
+            sandbox_kwargs["volumes"] = {self.volume_mount_path: self._volume}
+
+        return driver_command, sandbox_kwargs
+
+    # ------------------------------------------------------------------
+    # Lifecycle
+    # ------------------------------------------------------------------
+
+    def start(self) -> None:
+        """Start the Modal sandbox and initialize the driver process."""
+        if self._sandbox is not None:
+            return
+
+        app = self._resolve_app()
+        driver_command, sandbox_kwargs = self._build_driver_command_and_sandbox_kwargs(
+            app=app
+        )
+
+        self._sandbox = modal.Sandbox.create(**sandbox_kwargs)
+        self._proc = self._sandbox.exec(
+            "python", "-u", "-c", driver_command, bufsize=1, timeout=self.timeout
+        )
+        self._stdin = self._proc.stdin
+        self._stdout_iter = iter(self._proc.stdout)
+        self._stderr_iter = iter(getattr(self._proc, "stderr", []))
+        self._start_stdout_reader()
+
+    async def astart(self) -> None:
+        """Start the Modal sandbox and initialize the driver process (async)."""
+        if self._sandbox is not None:
+            return
+
+        app = await self._aresolve_app()
+        driver_command, sandbox_kwargs = self._build_driver_command_and_sandbox_kwargs(
+            app=app
+        )
+
+        self._sandbox = await modal.Sandbox.create.aio(**sandbox_kwargs)
+        self._proc = await self._sandbox.exec.aio(
+            "python", "-u", "-c", driver_command, bufsize=1, timeout=self.timeout
+        )
+        self._stdin = self._proc.stdin
+        self._stdout_iter = iter(self._proc.stdout)
+        self._stderr_iter = iter(getattr(self._proc, "stderr", []))
+        self._start_stdout_reader()
+
+    def shutdown(self) -> None:
+        """Terminate the sandbox and clean up all resources."""
+        if self._sandbox is not None:
+            try:
+                self._sandbox.terminate()
+            except Exception:
+                logger.exception(
+                    "Error while terminating Modal sandbox during shutdown"
+                )
+
+        self._sandbox = None
+        self._proc = None
+        self._stdin = None
+        self._stdout_iter = None
+        self._stdout_queue = None
+        self._stdout_reader_thread = None
+        self._stderr_iter = None
+        self._volume = None
+        with self._sub_lm_executor_lock:
+            if self._sub_lm_executor is not None:
+                self._sub_lm_executor.shutdown(wait=False, cancel_futures=True)
+                self._sub_lm_executor = None
+
+    async def ashutdown(self) -> None:
+        """Terminate the sandbox and clean up all resources (async)."""
+        if self._sandbox is not None:
+            try:
+                if hasattr(self._sandbox.terminate, "aio"):
+                    await self._sandbox.terminate.aio()
+                else:
+                    self._sandbox.terminate()
+            except Exception:
+                logger.exception(
+                    "Error while terminating Modal sandbox during async shutdown"
+                )
+
+        if self._stdout_reader_thread is not None:
+            self._stdout_reader_thread.join(timeout=2.0)
+            self._stdout_reader_thread = None
+
+        self._sandbox = None
+        self._proc = None
+        self._stdin = None
+        self._stdout_iter = None
+        self._stdout_queue = None
+        self._stderr_iter = None
+        self._volume = None
+        with self._sub_lm_executor_lock:
+            if self._sub_lm_executor is not None:
+                self._sub_lm_executor.shutdown(wait=False, cancel_futures=True)
+                self._sub_lm_executor = None
+
+    # ------------------------------------------------------------------
+    # Execution helpers
+    # ------------------------------------------------------------------
+
+    def _tool_names(self) -> list[str]:
+        """Get the list of registered tool names."""
+        tool_names = ["llm_query", "llm_query_batched"]
+        if self._tools:
+            tool_names.extend(self._tools.keys())
+        return tool_names
+
+    def _output_names(self) -> list[str]:
+        """Get the list of output field names."""
+        if not self.output_fields:
+            return []
+        return [
+            field["name"]
+            for field in self.output_fields
+            if isinstance(field, dict) and field.get("name")
+        ]
+
+    def _summarize_stdout(self, stdout: str) -> str:
+        """Summarize stdout output to prevent context window pollution."""
+        if not self.summarize_stdout:
+            return stdout
+        return _summarize_stdout(
+            stdout,
+            threshold=self.stdout_summary_threshold,
+            prefix_len=self.stdout_summary_prefix_len,
+        )
 
     @staticmethod
     def _is_recoverable_exec_channel_error(exc: Exception) -> bool:
         """Return ``True`` when an exec transport error is likely transient."""
-        return _is_recoverable_exec_channel_error_impl(exc)
+        text = str(exc).lower()
+        recoverable_signatures = (
+            "failed to write to exec stdin",
+            "broken pipe",
+            "sandbox input stream is not initialized",
+            "sandbox process exited unexpectedly",
+        )
+        return any(sig in text for sig in recoverable_signatures)
 
     @staticmethod
     def _is_recoverable_start_error(exc: Exception) -> bool:
         """Return ``True`` when sandbox startup failures are likely transient."""
-        return _is_recoverable_start_error_impl(exc)
+        text = str(exc).lower()
+        recoverable_signatures = (
+            "timed out",
+            "temporarily unavailable",
+            "connection reset",
+            "connection aborted",
+            "service unavailable",
+            "rate limit",
+            "429",
+            "resource exhausted",
+        )
+        return any(sig in text for sig in recoverable_signatures)
 
+    @contextmanager
     def execution_profile(self, profile: ExecutionProfile):
         """Temporarily override the default execution profile."""
-        return _execution_profile_impl(self, profile)
+        with execution_profile_context(self, profile) as current:
+            yield current
+
+    # ------------------------------------------------------------------
+    # Delegation
+    # ------------------------------------------------------------------
 
     def build_delegate_child(self, *, remaining_llm_budget: int) -> ModalInterpreter:
         """Build a child interpreter for recursive RLM delegation."""
@@ -337,6 +578,10 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
         )
         return child
 
+    # ------------------------------------------------------------------
+    # Execute
+    # ------------------------------------------------------------------
+
     def execute(
         self,
         code: str,
@@ -345,11 +590,222 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
         execution_profile: ExecutionProfile | None = None,
     ) -> str | FinalOutput:
         """Execute Python code in the Modal sandbox."""
-        return _execute_impl(
+        safe_vars: dict[str, Any] = {}
+        for key, value in (variables or {}).items():
+            try:
+                json.dumps(value)
+                safe_vars[key] = value
+            except TypeError:
+                safe_vars[key] = str(value)
+
+        request_payload = {
+            "code": code,
+            "variables": safe_vars,
+            "tool_names": self._tool_names(),
+            "output_names": self._output_names(),
+            "execution_profile": (
+                execution_profile or self.default_execution_profile
+            ).value,
+        }
+        profile_value = str(request_payload["execution_profile"])
+        code_hash, code_preview = summarize_code(code)
+        started_at = time.time()
+        emit_execution_event(
             self,
-            code,
-            variables,
-            execution_profile=execution_profile,
+            start_event_data(
+                execution_profile=profile_value,
+                code_hash=code_hash,
+                code_preview=code_preview,
+            ),
+        )
+
+        max_attempts = 3
+        for attempt in range(max_attempts):
+            if self._sandbox is None:
+                try:
+                    self.start()
+                except Exception as exc:
+                    can_retry = attempt < (max_attempts - 1) and (
+                        self._is_recoverable_start_error(exc)
+                    )
+                    if can_retry:
+                        self.shutdown()
+                        time.sleep(0.25 * (2**attempt))
+                        continue
+                    raise CodeInterpreterError(
+                        f"[sandbox_unavailable] Failed to start sandbox: {exc}"
+                    ) from exc
+
+            try:
+                self._write_line(request_payload)
+                deadline = (
+                    time.monotonic() + self.execute_timeout
+                    if self.execute_timeout
+                    else None
+                )
+
+                while True:
+                    remaining = (
+                        None
+                        if deadline is None
+                        else max(0.0, deadline - time.monotonic())
+                    )
+                    if remaining is not None and remaining <= 0:
+                        self.shutdown()
+                        raise CodeInterpreterError(
+                            f"Timed out waiting for sandbox response after {self.execute_timeout}s"
+                        )
+
+                    try:
+                        if self._stdout_queue is None:
+                            raise CodeInterpreterError(
+                                "Sandbox output queue is not initialized"
+                            )
+                        line = self._stdout_queue.get(timeout=remaining)
+                    except queue.Empty as exc:
+                        self.shutdown()
+                        raise CodeInterpreterError(
+                            f"Timed out waiting for sandbox response after {self.execute_timeout}s"
+                        ) from exc
+
+                    if line is None:
+                        stderr_tail = ""
+                        try:
+                            if self._stderr_iter is not None:
+                                stderr_tail = "".join(list(self._stderr_iter)[:50])
+                        except Exception:
+                            stderr_tail = ""
+                        msg = "Modal sandbox process exited unexpectedly."
+                        if stderr_tail:
+                            msg += f"\nStderr: {_redact_sensitive_text(stderr_tail)}"
+                        raise CodeInterpreterError(msg)
+
+                    try:
+                        message = json.loads(line)
+                    except json.JSONDecodeError:
+                        continue
+
+                    if "tool_call" in message:
+                        call = message["tool_call"] or {}
+                        name = call.get("name")
+                        args = call.get("args") or []
+                        kwargs = call.get("kwargs") or {}
+                        try:
+                            if name == "llm_query":
+                                result = self.llm_query(*args, **kwargs)
+                            elif name == "llm_query_batched":
+                                result = self.llm_query_batched(*args, **kwargs)
+                            elif name and name in self._tools:
+                                result = self._tools[name](*args, **kwargs)
+                            else:
+                                raise CodeInterpreterError(f"Unknown tool: {name}")
+
+                            try:
+                                json.dumps(result)
+                                reply = {"tool_result": result}
+                            except TypeError:
+                                reply = {"tool_result": str(result)}
+                        except Exception as exc:
+                            reply = {"tool_error": f"{type(exc).__name__}: {exc}"}
+
+                        self._write_line(reply)
+                        continue
+
+                    if "stdout" in message or "stderr" in message or "final" in message:
+                        stdout = message.get("stdout", "") or ""
+                        stderr = message.get("stderr", "") or ""
+                        final_obj = message.get("final")
+
+                        if final_obj is not None:
+                            output_keys = (
+                                list(final_obj.keys())[:50]
+                                if isinstance(final_obj, dict)
+                                else None
+                            )
+                            emit_execution_event(
+                                self,
+                                complete_event_data(
+                                    started_at=started_at,
+                                    execution_profile=profile_value,
+                                    code_hash=code_hash,
+                                    code_preview=code_preview,
+                                    success=True,
+                                    result_kind="final_output",
+                                    output_keys=output_keys,
+                                ),
+                            )
+                            return FinalOutput(final_obj)
+
+                        if stderr:
+                            summarized_stdout = self._summarize_stdout(stdout)
+                            emit_execution_event(
+                                self,
+                                complete_event_data(
+                                    started_at=started_at,
+                                    execution_profile=profile_value,
+                                    code_hash=code_hash,
+                                    code_preview=code_preview,
+                                    success=False,
+                                    result_kind="stderr",
+                                    stderr_preview=_redact_sensitive_text(stderr),
+                                ),
+                            )
+                            return (
+                                summarized_stdout
+                                + ("\n" if summarized_stdout else "")
+                                + stderr
+                            )
+
+                        stdout_preview = self._summarize_stdout(stdout)
+                        emit_execution_event(
+                            self,
+                            complete_event_data(
+                                started_at=started_at,
+                                execution_profile=profile_value,
+                                code_hash=code_hash,
+                                code_preview=code_preview,
+                                success=True,
+                                result_kind="stdout",
+                                stdout_preview=stdout_preview,
+                            ),
+                        )
+                        return stdout_preview
+            except Exception as exc:
+                can_retry = attempt < (max_attempts - 1) and (
+                    self._is_recoverable_exec_channel_error(exc)
+                )
+                if can_retry:
+                    self.shutdown()
+                    time.sleep(0.25 * (2**attempt))
+                    continue
+                emit_execution_event(
+                    self,
+                    complete_event_data(
+                        started_at=started_at,
+                        execution_profile=profile_value,
+                        code_hash=code_hash,
+                        code_preview=code_preview,
+                        success=False,
+                        result_kind="exception",
+                        error_type=type(exc).__name__,
+                        error=str(exc),
+                    ),
+                )
+                raise
+
+        emit_execution_event(
+            self,
+            complete_event_data(
+                started_at=started_at,
+                execution_profile=profile_value,
+                code_hash=code_hash,
+                code_preview=code_preview,
+                success=False,
+                result_kind="retry_exhausted",
+            ),
+        )
+        raise CodeInterpreterError(
+            "[sandbox_unavailable] Unexpected execute retry exhaustion"
         )
 
     async def aexecute(
@@ -369,13 +825,9 @@ class ModalInterpreter(LLMQueryMixin, VolumeOpsMixin):
             )
         return self.execute(code, variables, execution_profile=execution_profile)
 
-    def shutdown(self) -> None:
-        """Terminate the sandbox and clean up all resources."""
-        _shutdown_impl(self)
-
-    async def ashutdown(self) -> None:
-        """Terminate the sandbox and clean up all resources (async)."""
-        await _ashutdown_impl(self)
+    # ------------------------------------------------------------------
+    # Context managers
+    # ------------------------------------------------------------------
 
     def __enter__(self) -> ModalInterpreter:
         """Start the interpreter and return it for use as a context manager."""
